question 36 complete

+++ src/main/java/com/lessons/hibernate/question_36/Product.java
@@ -24,6 +24,6 @@ public class Product {
     private Integer salary;
 
     @Column(name = "age")
-    private Byte age;
+    private Short age;
 
 }
